requirements.txt
Django==2.1.5
django-ckeditor==5.4.0
django-cleanup==2.1.0
django-crispy-forms==1.8.1
django-js-asset==1.0.0
django-model-utils==4.0.0
django-notifications-hq==1.6.0
unidecode
Pillow
docutils